rewrite apiconstraints and routing structure to support accept header versioning
