--- app/services/api_constraints_service.rb
@@ -1,11 +1,0 @@
-class ApiConstraintsService
-  def initialize(options)
-    @version = options[:version]
-    @default = options[:default]
-  end
-
-  def matches?(request)
-    @default || request.headers['Accept'].
-                  include?("application/vnd.mybucket.v#{@version}")
-  end
-end
